+++ totalRP3/Locales/Locale.lua
@@ -1496,7 +1496,33 @@ If you wish to report %s's profile and you cannot target them you will need to o
 	DEFAULT_GUILD_RANK = "Member",
 
 	CO_GENERAL_DISABLE_WELCOME_MESSAGE = "Disable welcome message",
-	CO_GENERAL_DISABLE_WELCOME_MESSAGE_HELP = "Disables the welcome message displayed in the chat frame on login."
+	CO_GENERAL_DISABLE_WELCOME_MESSAGE_HELP = "Disables the welcome message displayed in the chat frame on login.",
+
+	WHATS_NEW_25_7 = [[# Changelog version 2.4.6
+
+## Added
+
+- Added support for 10.0.7 furbolg language.
+- Added voice reference, custom guild name and rank presets in additional info.
+  - Voice reference will be shown in the tooltip if filled in.
+  - Custom guild name/rank will replace the original fields in TRP tooltips.
+  - You may disable either of these changes via tooltip settings ("Show voice reference" and "Show custom guild names").
+- Added nameplates settings:
+  - Name and title max length on nameplates
+  - Display first name only
+  - Display custom guild name
+  - Disable customizations in instances
+  - Disable customizations on NPCs
+- IC/OOC fields from other players' profile are now selectable to make it easier to copy URLs.
+- Added setting to disable the welcome message on login.
+
+## Fixed
+
+- Fixed an issue with "Unknown" names potentially showing on tooltips after a loading screen.
+- Fixed an issue with the KuiNameplates module to properly fade/hide nameplates when appropriate.
+- Fixed an issue where the target frame would stay hidden after a UI reload with a target selected.
+
+]],
 };
 
 -- Use Ellyb to generate the Localization system